move fixtures and payload test to feature
