+++ tests/Feature/Fixtures/FakeListenerWithProperties.php
@@ -1,11 +1,10 @@
 <?php
 
-namespace Laravel\Horizon\Tests\Unit\Fixtures;
+namespace Laravel\Horizon\Tests\Feature\Fixtures;
 
 use Illuminate\Contracts\Events\Dispatcher;
-use Laravel\Horizon\Contracts\Silenced;
 
-class FakeListenerSilenced implements Silenced
+class FakeListenerWithProperties
 {
     /**
      * @var Dispatcher
